feat: adjusted error validations and via cep response interface
