+++ src/common/interfaces/via-cep-response.interface.ts
@@ -1,0 +1,8 @@
+export interface ViaCepResponseProps {
+  cep: string;
+  logradouro: string;
+  complemento: string;
+  bairro: string;
+  localidade: string;
+  estado: string;
+}
