+++ src/modules/stores/services/address.service.ts
@@ -1,15 +1,8 @@
 import { Injectable, HttpStatus, HttpException } from "@nestjs/common";
 import { LoggerService } from "src/common/logger/logger.service";
 import axios from "axios";
-
-export interface ViaCepResponseProps {
-  cep: string;
-  logradouro: string;
-  complemento: string;
-  bairro: string;
-  localidade: string;
-  estado: string;
-}
+import { validatePostalCode } from "src/utils/validate-postal-code.util";
+import { ViaCepResponseProps } from "src/common/interfaces/via-cep-response.interface";
 
 @Injectable()
 export class AddressService {
@@ -23,13 +16,16 @@ export class AddressService {
     this.logger.log(`Requesting address from ViaCep API through postal code: ${postalCode}`, correlationId);
 
     try {
+      this.logger.log(`Validating postal code: ${postalCode}`, correlationId);
+      await validatePostalCode(postalCode);
+
       const response = await axios.get<ViaCepResponseProps>(`https://viacep.com.br/ws/${postalCode}/json/`);
 
       this.logger.log(`Address successfully retrieved from ViaCep API. Response data: ${JSON.stringify(response.data)}`, correlationId);
       return response.data;
 
     } catch (error) {
-      this.logger.error(`Error requesting address from ViaCep API: ${error.message}`, correlationId);
+      this.logger.error(`Error requesting address from ViaCep API: ${error.message}`, error.stack, correlationId);
       throw new HttpException(`Error requesting address from ViaCep API: ${error.message}`, HttpStatus.BAD_REQUEST);
     }
   }
